% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsQuiz.R
\name{generateQuestion}
\alias{generateQuestion}
\title{Generate a question list}
\usage{
generateQuestion(question, answers, correct_answer)
}
\arguments{
\item{question}{the text to display for the question}

\item{answers}{a vector containing the multiple choice options}

\item{correct_answer}{a integer indicating the correct answer}
}
\description{
Generate a question list from some inputs
}
