% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsquiz-package.R
\docType{package}
\name{jsquiz-package}
\alias{jsquiz}
\alias{jsquiz-package}
\title{jsquiz: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
