% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsQuiz.R
\name{jsQuiz}
\alias{jsQuiz}
\title{A simple HTML widget for a multiple choice quiz in Javascript}
\usage{
jsQuiz(quizText, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{quizText}{a list of questions generated with \code{generateQuestion}}
}
\description{
Generate a multiple choice quiz from a list of questions and answers
}
